refactorizado para traajar con enteros mas grandes creado metodo main

+++ src/main/java/dev/gruhart/factorial/CalculaFactorial.java
@@ -2,7 +2,7 @@ package dev.gruhart.factorial;
 
 public class CalculaFactorial {
 
-    public int calcula(int i) {
+    public long calcula(long i) {
 
         while(i > 1){
             return i * calcula(i-1);
